+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Prevent `alchemy-select` from keeping on loading remote data when last few results were empty
 * Fix pagination breaking re-opening of `alchemy-select` dropdown
 * Fix initial value of a `BelongsTo` `alchemy-select` element not loading
+* Fix `alchemy-select` downward-arrow-icon being rendered over value content
 
 ## 0.1.7 (2022-06-12)