+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix pagination breaking re-opening of `alchemy-select` dropdown
 * Fix initial value of a `BelongsTo` `alchemy-select` element not loading
 * Fix `alchemy-select` downward-arrow-icon being rendered over value content
+* Use `alchemy-code-input` for HTML fields 
 
 ## 0.1.7 (2022-06-12)