+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.8 (WIP)
+## 0.1.8 (2022-06-29)
 
 * Move the `alchemy-select` pagination checker initializer to a static function
 * Prevent `alchemy-select` from keeping on loading remote data when last few results were empty